fix some routing issues verkiezingsuitslag
