+++ app/routes/verkiezingen/verkiezingsuitslag/prepare.js
@@ -1,21 +1,9 @@
 import Route from '@ember/routing/route';
-import { inject as service } from '@ember/service';
 
 export default class PrepareInstallatievergaderingRoute extends Route {
-  @service store;
-
   async model() {
-    const parentModel = this.modelFor('verkiezingen.verkiezingsuitslag');
-    const installatievergadering = parentModel.installatievergadering;
-    const ivStatuses = await this.store.findAll(
-      'installatievergadering-status'
-    );
-    const selectedStatus = installatievergadering.get('status');
-
     return {
-      installatievergadering,
-      ivStatuses,
-      selectedStatus,
+      ...this.modelFor('verkiezingen.verkiezingsuitslag'),
     };
   }
 }
